spring-jpa-hibernate/src/test/java/dev/intermediatebox/flat/repository/CourseRepositoryTest.java
package dev.intermediatebox.flat.repository;

import dev.intermediatebox.JpaHibernateApplication;
import dev.intermediatebox.flat.entity.Course;
import lombok.extern.slf4j.Slf4j;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.DirtiesContext;

import javax.persistence.EntityManager;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNull;

@Slf4j
@SpringBootTest(classes = JpaHibernateApplication.class)
public class CourseRepositoryTest {
	@Autowired
  CourseRepository repository;

	@Autowired
	EntityManager em;

	@Test
	public void findById_basic() {
		Course course = repository.findById(10001L);
		assertEquals("course 1", course.getName());
	}

	@Test
	@DirtiesContext
	public void deleteById_basic() {
		repository.deleteById(10002L);
		assertNull(repository.findById(10002L));
	}

	@Test
	@DirtiesContext
	public void save_basic() {
		// get a course
		Course course = repository.findById(10001L);
		assertEquals("course 1", course.getName());

		// update details
		course.setName("course 1 - Updated");
		repository.save(course);

		// check the value
		Course course1 = repository.findById(10001L);
		assertEquals("course 1 - Updated", course1.getName());
	}
}
